requirements.txt
langchain
langchain-openai
celery
python-dotenv
fastapi
uvicorn
requests

